backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Clément MOINE <clement.moine86@gmail.com>"]

[tool.poetry.dependencies]
requests = "^2.31.0"
colorama = "^0.4.6"
beautifulsoup4 = "^4.12.3"
python = "^3.12"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
unidecode = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
